+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# v0.14.3 (unreleased)
+* Follow 301 and 302 responses to POST with a GET request.
+* Follow all redirect responses to HEAD with a HEAD request.
+* Fix bug where redirect response is disposed prematurely even if not followed.
+* Fix bug where payload headers are sent from original request when
+  following a redirect with a GET or HEAD.
+* Return a failed task instead of throwing when a client callback
+  throws an exception. Fixes a resource leak.
+
 # v0.14.2 (2016-08-10)
 * Override `getMessage` in `UnexpectedStatus`